pyproject.toml
[project]
name = "ragnews"
version = "0.1.0"
description = "RAGNews is a news aggregation platform that uses RAG to aggregate news from various sources."
license-files = ["LICENSE"]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
  "annotated-types==0.7.0",
  "anyio==4.9.0",
  "async-timeout==4.0.3",
  "certifi==2025.4.26",
  "cfgv==3.4.0",
  "charset-normalizer==3.4.2",
  "distlib==0.3.9",
  "exceptiongroup==1.3.0",
  # "faiss-gpu==1.7.2",
  "filelock==3.18.0",
  "greenlet==3.2.2",
  "h11==0.16.0",
  "httpcore==1.0.9",
  "httpx==0.28.1",
  "identify==2.6.10",
  "idna==3.10",
  "jiter==0.10.0",
  "jsonpatch==1.33",
  "jsonpointer==3.0.0",
  "langchain==0.3.25",
  "langchain-core==0.3.69",
  "langchain-openai==0.3.28",
  "langchain-text-splitters==0.3.8",
  "langsmith==0.4.8",
  "loguru==0.7.3",
  "nodeenv==1.9.1",
  "numpy==1.26.4",
  "openai==1.97.0",
  "orjson==3.11.0",
  "packaging==25.0",
  "platformdirs==4.3.8",
  "pre-commit==4.2.0",
  "pydantic==2.11.4",
  "pydantic-core==2.33.2",
  "pyyaml==6.0.2",
  "regex==2024.11.6",
  "requests==2.32.4",
  "requests-toolbelt==1.0.0",
  "sniffio==1.3.1",
  "sqlalchemy==2.0.41",
  "streamlit==1.39.0",
  "tenacity==9.1.2",
  "tiktoken==0.9.0",
  "tqdm==4.67.1",
  "typing-extensions==4.13.2",
  "typing-inspection==0.4.0",
  "urllib3==2.4.0",
  "virtualenv==20.31.2",
  "zstandard==0.23.0",
  "crawl4ai==0.7.1"
]

[project.optional-dependencies]
dev = ["pytest", "black"]

[tool.setuptools]
packages = ["rag_news"]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
